+++ src/app/system/configurations/global-configurations-tab/clone-configuration/clone-configuration.component.ts
@@ -10,6 +10,11 @@ import { SystemService } from 'app/system/system.service';
   styleUrls: ['./clone-configuration.component.scss']
 })
 export class CloneConfigurationComponent implements OnInit {
+
+  private readonly configLevels: Record<string, any> = {
+    "payment-channel": "payment-channel-ou-level"
+  };
+
   /** Global Configuration form. */
   configurationForm: UntypedFormGroup;
   /** Configuration. */
@@ -18,6 +23,8 @@ export class CloneConfigurationComponent implements OnInit {
   listCountries: any = [];
   listCountriesSliced: any = [];
   configId: number;
+  offices: any[] = [];
+
   /**
    * Retrieves the configuration data from `resolve`.
    * @param {FormBuilder} formBuilder Form Builder.
@@ -50,11 +57,22 @@ export class CloneConfigurationComponent implements OnInit {
    * Creates and sets the global configuration form.
    */
   createConfigurationForm () {
+    let countryId = null;
+    // Set the initial value for countryId if cloning a configuration with an office
+    if (this.configuration.officeId) {
+      countryId = this.configuration.country.id;
+    }
     this.configurationForm = this.formBuilder.group({
-      value: ['', Validators.required],
-      countryId: [null, Validators.required],
-      globalConfigId: [this.configId]
+      value: ['', !this.configuration.codeId ? Validators.required : null],
+      countryId: [countryId, Validators.required],
+      globalConfigId: [this.configId],
+      codeValueId: [this.configuration.codeValueId, this.configuration.codeId ? Validators.required : null],
+      officeId: [null]
     });
+    // Populate the offices dropdown if countryId has been set above
+    if (countryId) {
+      this.onCountryChange(countryId);
+    }
   }
 
   getCountries () {
@@ -78,5 +96,23 @@ export class CloneConfigurationComponent implements OnInit {
       .subscribe((response: any) => {
         this.router.navigate(['../../'], { relativeTo: this.route });
       });
+  }
+
+  /**
+   * Populates the offices dropdown based on the selected country.
+   * @param countryId Selected country ID.
+   * @returns void
+   */
+  onCountryChange(countryId: number) {
+    const levelConfig = this.configLevels[this.configuration.name];
+    if (!levelConfig) return;
+
+    // Clear existing office selection if the country changes
+    if (this.configuration.country?.id != countryId) {
+      this.configurationForm.get('officeId').reset();
+    }
+    this.organizationService.searchOffices({countryId, levelConfig}).subscribe(res => {
+      this.offices = res;
+    });
   }
 }